requirements.txt
scikit-learn==0.22.2.post1
pyAudioAnalysis==0.3.14
speechemotionrecognition @ git+https://github.com/allenanhu/speech-emotion-recognition
numpy==1.19.5
pliers==0.4.2
pydub==0.25.1
pillow==6.1.0
keras==2.2.5
imageio==2.3
speechpy==2.2
imbalanced-learn==0.6.2
pandas==1.3.5
scipy==1.4.1
tensorflow==1.15.0
grpcio==1.62.3
grpcio-status==1.48.2
h5py==2.10.0
protobuf==3.20.0
moviepy==0.2.3.5
google-api-core==2.8.2
google-cloud-core==2.4.1
google-cloud-speech==2.15.1
google-cloud-storage==2.14.0
googleapis-common-protos==1.56.4
eyed3==0.9.7
matplotlib==3.2.2
plotly==5.18.0
ipykernel==6.16.2
python-dotenv==0.21.1
tqdm==4.66.5